chore(ERTP): AmountMath compare should be same as Key compare'
