+++ packages/ERTP/package.json
@@ -48,11 +48,11 @@
     "@endo/far": "^0.2.18",
     "@endo/marshal": "^0.8.5",
     "@endo/nat": "^4.1.27",
-    "@endo/promise-kit": "^0.2.56"
+    "@endo/promise-kit": "^0.2.56",
+    "@fast-check/ava": "^1.1.3"
   },
   "devDependencies": {
     "@endo/bundle-source": "^2.5.1",
-    "@fast-check/ava": "^1.1.3",
     "ava": "^5.2.0",
     "tsd": "^0.28.1"
   },
@@ -67,7 +67,8 @@
       "# fast-check unsupported",
       "test/unitTests/test-amountProperties.js",
       "test/unitTests/test-inputValidation.js",
-      "test/unitTests/test-issuerObj.js"
+      "test/unitTests/test-issuerObj.js",
+      "test/unitTests/test-amount-key-arith.js"
     ]
   },
   "ava": {
