+++ packages/ERTP/test/unitTests/mathHelpers/test-copyBagMathHelpers.js
@@ -6,7 +6,7 @@ import {
 } from '@agoric/store';
 
 import { AmountMath as m, AssetKind } from '../../../src/index.js';
-import { mockBrand } from './mockBrand.js';
+import { mockBrand } from '../../../tools/mockBrand.js';
 
 // The "unit tests" for MathHelpers actually make the calls through
 // AmountMath so that we can test that any duplication is handled
