+++ packages/ERTP/test/unitTests/test-amountProperties.js
@@ -1,29 +1,9 @@
 import { test } from '@agoric/swingset-vat/tools/prepare-test-env-ava.js';
-import { makeCopyBag } from '@agoric/store';
 import { fc } from '@fast-check/ava';
 
 import { AmountMath as m, AssetKind } from '../../src/index.js';
-import { mockBrand } from './mathHelpers/mockBrand.js';
-
-// Perhaps makeCopyBag should coalesce duplicate labels, but for now, it does
-// not.
-const distinctLabels = pairs =>
-  new Set(pairs.map(([label, _qty]) => label)).size === pairs.length;
-const positiveCounts = pairs =>
-  pairs.filter(([_l, qty]) => qty > 0n).length === pairs.length;
-const arbBagContents = fc
-  .nat(7)
-  .chain(size =>
-    fc.array(
-      fc.tuple(fc.string(), fc.bigUint({ max: 1_000_000_000_000_000n })),
-      { minLength: size, maxLength: size },
-    ),
-  )
-  .filter(pairs => distinctLabels(pairs) && positiveCounts(pairs));
-
-const arbAmount = arbBagContents.map(contents =>
-  m.make(mockBrand, harden(makeCopyBag(contents))),
-);
+import { mockBrand } from '../../tools/mockBrand.js';
+import { arbAmount } from '../../tools/arb-amount.js';
 
 // Note: we write P => Q as !P || Q since JS has no logical => operator
 const implies = (p, q) => !p || q;
