+++ packages/ERTP/tools/mockBrand.js
@@ -1,10 +1,10 @@
 import { Far } from '@endo/marshal';
-import { AssetKind } from '../../../src/index.js';
+import { AssetKind } from '../src/index.js';
 
 /** @type {Brand<AssetKind>} */
 export const mockBrand = Far('brand', {
   // eslint-disable-next-line no-unused-vars
-  isMyIssuer: async allegedIssuer => false,
+  isMyIssuer: async _allegedIssuer => false,
   getAllegedName: () => 'mock',
   getAmountShape: () => {},
   getDisplayInfo: () => ({
